requirements.txt
annotated-types==0.7.0 ; python_version >= "3.10"
anyio==4.10.0 ; python_version >= "3.10"
async-timeout==5.0.1 ; python_version >= "3.10" and python_full_version < "3.11.3"
certifi==2025.8.3 ; python_version >= "3.10"
click==8.2.1 ; python_version >= "3.10"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_version >= "3.10"
exceptiongroup==1.3.0 ; python_version == "3.10"
fastapi==0.116.1 ; python_version >= "3.10"
h11==0.16.0 ; python_version >= "3.10"
httpcore==1.0.9 ; python_version >= "3.10"
httptools==0.6.4 ; python_version >= "3.10"
httpx==0.27.2 ; python_version >= "3.10"
idna==3.10 ; python_version >= "3.10"
loguru==0.7.2 ; python_version >= "3.10"
orjson==3.11.3 ; python_version >= "3.10"
pydantic-core==2.33.2 ; python_version >= "3.10"
pydantic-settings==2.10.1 ; python_version >= "3.10"
pydantic==2.11.7 ; python_version >= "3.10"
python-dotenv==1.1.1 ; python_version >= "3.10"
pyyaml==6.0.2 ; python_version >= "3.10"
redis==6.4.0 ; python_version >= "3.10"
sniffio==1.3.1 ; python_version >= "3.10"
starlette==0.47.3 ; python_version >= "3.10"
typing-extensions==4.15.0 ; python_version >= "3.10"
typing-inspection==0.4.1 ; python_version >= "3.10"
uvicorn==0.35.0 ; python_version >= "3.10"
uvloop==0.21.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.10"
watchfiles==1.1.0 ; python_version >= "3.10"
websockets==15.0.1 ; python_version >= "3.10"
win32-setctime==1.2.0 ; python_version >= "3.10" and sys_platform == "win32"
pandas>=2.2
pandas-datareader>=0.10
yfinance>=0.2.27
fredapi>=0.4.3
pyyaml>=6.0
